Extract yaml parsing from collection
This will make it easier to make the yaml able to do determine different
things to be done when the test is run
